cmd: compute costs in background
This commit changes the project to compute costs in the background and
store the resulting metrics in a cache. Metrics are then served from
this cache to avoid unnecessarily long computations during scraping.

This change has the added benefit of allowing the removal of time limits
on the database query, since they entire processing occurs in the
background.

Fixes: #2

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>